requirements.txt
httpx[http2]
lockfile
fastapi
pillow
pydantic-settings
uvicorn[standard]